move status mapping to Bridge::Utility
