+++ library/common/bridge/BUILD
@@ -12,6 +12,8 @@ envoy_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//library/common/types:c_types_lib",
-        "@envoy_mobile//library/common/data:utility_lib",
+        "//library/common/data:utility_lib",
+        "@envoy//envoy/http:codes_interface",
+        "@envoy//source/common/http:codes_lib",
     ],
 )
