+++ library/common/bridge/utility.cc
@@ -1,9 +1,25 @@
+#include "library/common/bridge/utility.h"
+
 #include <string>
 
 #include "library/common/data/utility.h"
 
 namespace Envoy {
 namespace Bridge {
+namespace Utility {
+
+envoy_error_code_t errorCodeFromLocalStatus(Http::Code status) {
+  switch (status) {
+  case Http::Code::RequestTimeout:
+    return ENVOY_REQUEST_TIMEOUT;
+  case Http::Code::PayloadTooLarge:
+    return ENVOY_BUFFER_LIMIT_EXCEEDED;
+  case Http::Code::ServiceUnavailable:
+    return ENVOY_CONNECTION_FAILURE;
+  default:
+    return ENVOY_UNDEFINED_ERROR;
+  }
+}
 
 envoy_map makeEnvoyMap(std::vector<std::pair<std::string, std::string>> pairs) {
   envoy_map_entry* fields =
@@ -23,5 +39,6 @@ envoy_map makeEnvoyMap(std::vector<std::pair<std::string, std::string>> pairs) {
   return new_event;
 }
 
+} // namespace Utility
 } // namespace Bridge
 } // namespace Envoy
