+++ library/common/extensions/filters/http/platform_bridge/BUILD
@@ -28,6 +28,7 @@ envoy_cc_extension(
     deps = [
         ":filter_cc_proto",
         "//library/common/api:external_api_lib",
+        "//library/common/bridge:utility_lib",
         "//library/common/data:utility_lib",
         "//library/common/http:header_utility_lib",
         "//library/common/http:internal_headers_lib",
